# backend/docs/swagger.yaml
basePath: /
definitions:
...
  http.CommentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/http.CommentResponse'
        type: array
      pagination:
        $ref: '#/definitions/http.CursorPagination'
    type: object
  http.CommentRequest:
    properties:
      content:
        type: string
      parent_id:
        type: string
    type: object
  http.CommentResponse:
    properties:
      author_id:
        type: string
      author_name:
        type: string
      book_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      depth:
        type: integer
      id:
        type: string
      parent_id:
        type: string
      reply_count:
        type: integer
      root_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  http.CursorPagination:
    properties:
      has_more:
        example: true
        type: boolean
      next_cursor:
        type: string
    type: object
...
  http.ModerateCommentRequest:
    properties:
      status:
        type: string
    type: object
...
  http.UpdateCommentRequest:
    properties:
      content:
        type: string
    type: object
...
info:
  contact: {}
  description: Public HTTP API exposed by the Book Store API Gateway.
  title: Book Store API
  version: "1.0"
paths:
...
  /api/v1/admin/comments/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: published or hidden
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ModerateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.CommentResponse'
      security:
      - BearerAuth: []
      summary: Hide or publish a comment
      tags:
      - Admin
      - Comments
...
  /api/v1/books/{id}/comments:
    get:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Opaque cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.CommentListResponse'
      summary: List root comments of a book
      tags:
      - Comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment content and optional parent_id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.CommentResponse'
      security:
      - BearerAuth: []
      summary: Comment on a book or reply to a comment
      tags:
      - Comments
...
  /api/v1/comments/{id}:
    delete:
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.CommentResponse'
      security:
      - BearerAuth: []
      summary: Soft-delete own comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: New content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.CommentResponse'
      security:
      - BearerAuth: []
      summary: Update own comment
      tags:
      - Comments
  /api/v1/comments/{id}/replies:
    get:
      parameters:
      - description: Root comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Opaque cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.CommentListResponse'
      summary: List replies in a comment thread
      tags:
      - Comments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token using the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"